+++ Tests/MoyaProviderSpec.swift
@@ -771,6 +771,48 @@ class MoyaProviderSpec: QuickSpec {
 
         }
 
+        describe("a provider with upload progress tracking") {
+            var provider: MoyaProvider<HTTPBin>!
+            beforeEach {
+                provider = MoyaProvider<HTTPBin>()
+            }
+
+            it("tracks progress of request") {
+
+                let url = Bundle(for: MoyaProviderSpec.self).urlForImageResource("testImage.png")!
+                let target: HTTPBin = .upload(file: url)
+                
+                var progressObjects: [Progress?] = []
+                var progressValues: [Double] = []
+                var completedValues: [Bool] = []
+                var error: MoyaError?
+
+                waitUntil(timeout: 5.0) { done in
+                    let progressClosure: ProgressBlock = { progress in
+                        progressObjects.append(progress.progressObject)
+                        progressValues.append(progress.progress)
+                        completedValues.append(progress.completed)
+                    }
+
+                    let progressCompletionClosure: Completion = { (result) in
+                        if case .failure(let err) = result {
+                            error = err
+                        }
+                        done()
+                    }
+
+                    provider.request(target, callbackQueue: nil, progress: progressClosure, completion: progressCompletionClosure)
+                }
+
+                expect(error).to(beNil())
+                expect(progressValues.count) > 3
+                expect(completedValues.count) > 3
+                expect(completedValues.filter{ !$0 }.count) == completedValues.count - 1 // only false except one
+                expect(completedValues.last) == true // the last must be true
+                expect(progressObjects.filter{$0 != nil}.count) == progressObjects.count // no nil object
+            }
+        }
+
         describe("using a custom callback queue") {
             var stubDescriptor: OHHTTPStubsDescriptor!